Add `Into<tonic::Status>` bound to `Snapshotter::Error`
This allows implementers to express which grpc status should be
returned to clients on errors.

Prior to this patch, all errors got converted to "internal"
errors (`tonic::Status::internal`), which doesn't work when
specific status are needed. For example, remote snapshotters
need to return "already exists" to indicate that a layer
doesn't need to be downloaded.

Signed-off-by: Wedson Almeida Filho <walmeida@microsoft.com>